Implement profile metadata routing and add coverage for wallet, formatting, and missing-job error paths.
This ensures profile pages render route-aware titles with invalid-address fallback while adding focused tests for wallet provider behavior, amount formatting rounding, and missing job ids.

Made-with: Cursor
